pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = "Discord bot for the PlaceTW discord server"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
supabase = "^2.4.6"
babel = "^2.15.0"
numpy = "^1.26.4"
requests = "^2.32.2"
pygithub = "^2.3.0"
detecto = "^1.2.2"
shortuuid = "^1.0.13"
validators = "^0.28.2"
beautifulsoup4 = "^4.12.3"
gitpython = "^3.1.43"
pytest = "^8.2.1"
discord-py = "^2.4.0"
pytest-asyncio = "^0.23.7"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true